Moved many pages and components to the common folder so we can share with WebApp.
