+++ src/js/common/components/Campaign/RecommendedCampaignListRetrieveController.jsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import AppObservableStore, { messageService } from '../../common/stores/AppObservableStore';
-import CampaignActions from '../../common/actions/CampaignActions';
-import initializejQuery from '../../common/utils/initializejQuery';
-import { renderLog } from '../../common/utils/logging';
-import VoterStore from '../../stores/VoterStore';
+import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
+import CampaignActions from '../../actions/CampaignActions';
+import initializejQuery from '../../utils/initializejQuery';
+import { renderLog } from '../../utils/logging';
+import VoterStore from '../../../stores/VoterStore';
 
 
 class RecommendedCampaignListRetrieveController extends Component {
