+++ src/js/common/components/Navigation/CampaignSupportSteps.jsx
@@ -2,12 +2,12 @@ import withStyles from '@mui/styles/withStyles';
 import { Done } from '@mui/icons-material';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import commonMuiStyles from '../../common/components/Style/commonMuiStyles';
-import { InnerWrapper, OuterWrapperPageTitle, OuterWrapperSteps, PageTitle, StepCircle, StepNumber, StepWrapper } from '../../common/components/Style/stepDisplayStyles';
-import historyPush from '../../common/utils/historyPush';
-import { renderLog } from '../../common/utils/logging';
-import AppObservableStore, { messageService } from '../../common/stores/AppObservableStore';
-import CampaignSupporterStore from '../../common/stores/CampaignSupporterStore';
+import commonMuiStyles from '../Style/commonMuiStyles';
+import { InnerWrapper, OuterWrapperPageTitle, OuterWrapperSteps, PageTitle, StepCircle, StepNumber, StepWrapper } from '../Style/stepDisplayStyles';
+import historyPush from '../../utils/historyPush';
+import { renderLog } from '../../utils/logging';
+import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
+import CampaignSupporterStore from '../../stores/CampaignSupporterStore';
 
 class CampaignSupportSteps extends Component {
   constructor (props) {
