+++ src/js/common/components/Style/CampaignProcessStyles.jsx
