+++ src/js/common/components/SuperSharing/AddContactsFromGoogle.jsx
@@ -3,11 +3,11 @@ import withStyles from '@mui/styles/withStyles';
 import { loadGapiInsideDOM } from 'gapi-script';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import VoterActions from '../../actions/VoterActions';
-import { renderLog } from '../../common/utils/logging';
-import webAppConfig from '../../config'; // eslint-disable-line import/no-cycle
+import VoterActions from '../../../actions/VoterActions';
+import { renderLog } from '../../utils/logging';
+import webAppConfig from '../../../config'; // eslint-disable-line import/no-cycle
 import AddContactConsts from '../../constants/AddContactConsts';
-import VoterStore from '../../stores/VoterStore';
+import VoterStore from '../../../stores/VoterStore';
 
 
 class AddContactsFromGoogle extends Component {
