+++ src/js/common/constants/AddContactConsts.js
