+++ src/js/common/pages/CampaignSupport/CampaignSupportPayToPromoteProcess.jsx
@@ -7,22 +7,22 @@ import { loadStripe } from '@stripe/stripe-js';
 import PropTypes from 'prop-types';
 import React, { Component, Suspense } from 'react';
 import Helmet from 'react-helmet';
-import CampaignActions from '../../common/actions/CampaignActions';
-import DonationListForm from '../../common/components/Donation/DonationListForm';
-import InjectedCheckoutForm from '../../common/components/Donation/InjectedCheckoutForm';
-import standardBoxShadow from '../../common/components/Style/standardBoxShadow';
-import { OuterWrapper, PageWrapper } from '../../common/components/Style/stepDisplayStyles';
-import LoadingWheelComp from '../../common/components/Widgets/LoadingWheelComp';
-import DonateStore from '../../common/stores/DonateStore';
-import historyPush from '../../common/utils/historyPush';
-import { renderLog } from '../../common/utils/logging';
-import { SkipForNowButtonPanel, SkipForNowButtonWrapper } from '../../common/components/Style/CampaignSupportStyles';
-import webAppConfig from '../../config';
-import AppObservableStore, { messageService } from '../../common/stores/AppObservableStore';
-import CampaignStore from '../../common/stores/CampaignStore';
-import VoterStore from '../../stores/VoterStore';
-import { getCampaignXValuesFromIdentifiers, retrieveCampaignXFromIdentifiersIfNeeded } from '../../common/utils/campaignUtils';
-import initializejQuery from '../../common/utils/initializejQuery';
+import CampaignActions from '../../actions/CampaignActions';
+import DonationListForm from '../../components/Donation/DonationListForm';
+import InjectedCheckoutForm from '../../components/Donation/InjectedCheckoutForm';
+import standardBoxShadow from '../../components/Style/standardBoxShadow';
+import { OuterWrapper, PageWrapper } from '../../components/Style/stepDisplayStyles';
+import LoadingWheelComp from '../../components/Widgets/LoadingWheelComp';
+import DonateStore from '../../stores/DonateStore';
+import historyPush from '../../utils/historyPush';
+import { renderLog } from '../../utils/logging';
+import { SkipForNowButtonPanel, SkipForNowButtonWrapper } from '../../components/Style/CampaignSupportStyles';
+import webAppConfig from '../../../config';
+import AppObservableStore, { messageService } from '../../stores/AppObservableStore';
+import CampaignStore from '../../stores/CampaignStore';
+import VoterStore from '../../../stores/VoterStore';
+import { getCampaignXValuesFromIdentifiers, retrieveCampaignXFromIdentifiersIfNeeded } from '../../utils/campaignUtils';
+import initializejQuery from '../../utils/initializejQuery';
 
 
 const stripePromise = loadStripe(webAppConfig.STRIPE_API_KEY);
