+++ src/js/common/utils/superSharingStepPaths.js
