+++ src/js/common/utils/superSharingSuggestedEmailText.js
