+++ src/js/pages/CampaignStart/CampaignStartAddPhoto.jsx
@@ -12,7 +12,7 @@ import historyPush from '../../common/utils/historyPush';
 import { renderLog } from '../../common/utils/logging';
 import CampaignPhotoUpload from '../../components/CampaignStart/CampaignPhotoUpload';
 import CampaignStartSteps from '../../components/Navigation/CampaignStartSteps';
-import { CampaignProcessStepIntroductionText, CampaignProcessStepTitle } from '../../components/Style/CampaignProcessStyles';
+import { CampaignProcessStepIntroductionText, CampaignProcessStepTitle } from '../../common/components/Style/CampaignProcessStyles';
 import { CampaignStartDesktopButtonPanel, CampaignStartDesktopButtonWrapper, CampaignStartMobileButtonPanel, CampaignStartMobileButtonWrapper, CampaignStartSection, CampaignStartSectionWrapper } from '../../components/Style/CampaignStartStyles';
 import AppObservableStore, { messageService } from '../../common/stores/AppObservableStore';
 import CampaignStartStore from '../../stores/CampaignStartStore';
