+++ src/js/pages/Membership.jsx
@@ -19,7 +19,7 @@ import initializejQuery from '../common/utils/initializejQuery';
 
 
 const stripePromise = loadStripe(webAppConfig.STRIPE_API_KEY);
-const VoterFirstRetrieveController = loadable(() => import(/* webpackChunkName: 'VoterFirstRetrieveController' */ '../components/Settings/VoterFirstRetrieveController'));
+const VoterFirstRetrieveController = loadable(() => import(/* webpackChunkName: 'VoterFirstRetrieveController' */ '../common/components/Settings/VoterFirstRetrieveController'));
 
 class Membership extends Component {
   constructor (props) {
